refactor:Structured intoComponent
