+++ Lecture4/HeaderComponent/HeaderComponent.js
@@ -1,0 +1,13 @@
+const HeaderComponent = () => (<nav className="navbar navbar-light bg-light">
+    <a className="navbar-brand" href="#">
+        <img
+            className="d-inline-block align-top mr-4"
+            width={50}
+            src="https://images.weserv.nl/?url=https%3A%2F%2Fdrive.google.com%2Fuc%3Fexport%3Dview%26id%3D1ZN0OHfI7ngyLNIn0mDY36iFE2AewNMJ3&w=290&h=250&fit=cover&q=100"
+            alt=""
+        />
+        Brainy Fools
+    </a>
+</nav>);
+
+export default HeaderComponent;
